Refactor into multiple files. Swap from setup.py to pyproject.toml. Finish implementing new DBconnect functions
