+++ pyproject.toml
@@ -1,0 +1,34 @@
+[project]
+name = "splunksecrets"
+version = "1.1.0"
+description = "Command-line tools for working with secrets for Splunk products"
+requires-python = ">=3.7"
+dependencies = [
+    "click>=8.0.0",
+    "cryptography>=43",
+    "pcrypt"
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.metadata]
+allow-direct-references = true
+
+[project.scripts]
+splunksecrets = "splunksecrets.cli:main"
+
+[tool.hatch.build.targets.wheel]
+packages = ["splunksecrets"]
+
+[project.optional-dependencies]
+test = [
+    "pytest>=7.0.0",
+    "pytest-asyncio>=0.21.0"
+]
+
+[tool.pytest.ini_options]
+asyncio_mode = "auto"
+testpaths = ["tests"]
+python_files = ["test_*.py"]
